Tas 24 User Tagging

+++ app/routers/posts.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, HTTPException, Path
 from starlette import status
-from db import db_dependency, Posts, Comments
+from db import db_dependency, Posts, Comments, Users
 from .users import user_dependency
 from schemes import (
     PostCreate,
@@ -10,7 +10,9 @@ from schemes import (
     ReactionListResponse,
     GetComments,
     GetReactions,
+    UserTag,
 )
+import json
 from typing import List
 from sqlalchemy.orm import joinedload
 
@@ -34,6 +36,7 @@ async def get_all_posts(db: db_dependency):
         PostResponse(
             id=post.id,
             created_by=post.created_by,
+            tagged_users=post.get_tagged_user(),
             content=post.content,
             image_url=post.image_url,
             created_at=post.created_at,
@@ -72,7 +75,7 @@ async def get_all_posts(db: db_dependency):
 
 
 @router.get(
-    "user/{user_id}",
+    "/user/{user_id}",
     status_code=status.HTTP_200_OK,
     response_model=List[PostResponse],
 )
@@ -88,6 +91,7 @@ async def get_user_timeline(db: db_dependency, user_id: int = Path(gt=0)):
         PostResponse(
             id=post.id,
             created_by=post.created_by,
+            tagged_users=post.get_tagged_user(),
             content=post.content,
             image_url=post.image_url,
             created_at=post.created_at,
@@ -126,30 +130,52 @@ async def get_user_timeline(db: db_dependency, user_id: int = Path(gt=0)):
 
 
 @router.post(
-    "/create", status_code=status.HTTP_201_CREATED, response_model=CreatePostResponse
+    "/create/", status_code=status.HTTP_201_CREATED, response_model=CreatePostResponse
 )
 async def create_post(
-    user: user_dependency, db: db_dependency, post_request: PostCreate
+    user: user_dependency,
+    db: db_dependency,
+    post_request: PostCreate,
 ):
     if user is None:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed"
         )
 
+    tagged_users = post_request.tagged_users
+
+    for username in tagged_users:
+        if tagged_users == user.get("username"):
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="User can't tag themself",
+            )
+
+        check_tagged_users = db.query(Users).filter(Users.username == username).first()
+        if check_tagged_users is None:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail="User to tag not found"
+            )
+
     post_model = Posts(
-        **post_request.model_dump(),
         created_by=user.get("username"),
-        owner_id=user.get("id")
+        owner_id=user.get("id"),
+        content=post_request.content,
+        image_url=post_request.image_url,
     )
+
+    post_model.set_tagged_user(tagged_users)
 
     db.add(post_model)
     db.commit()
+    db.refresh(post_model)
 
     return {
         "detail": "Post created successfully",
         "post_details": PostResponse(
             id=user.get("id"),
             created_by=user.get("username"),
+            tagged_users=post_model.get_tagged_user(),
             content=post_model.content,
             image_url=post_model.image_url,
             created_at=post_model.created_at,
@@ -193,6 +219,109 @@ async def update_post(
     return {"detail": "Post updated successfully"}
 
 
+@router.post(
+    "/{post_id}/add_user_tag",
+    status_code=status.HTTP_200_OK,
+)
+async def add_user_tag(
+    user: user_dependency,
+    db: db_dependency,
+    user_tag: UserTag,
+    post_id: int = Path(gt=0),
+):
+    if user is None:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed"
+        )
+
+    query_post = db.query(Posts).filter(Posts.id == post_id).first()
+
+    if query_post is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+        )
+
+    post = db.query(Posts).filter(Posts.owner_id == user.get("id")).first()
+
+    if post is None:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Not authorized to update this post",
+        )
+
+    if query_post.tagged_user:
+        try:
+            existing_tags = json.loads(query_post.tagged_user)
+            if existing_tags:
+                raise HTTPException(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail="Post already has user tags",
+                )
+
+        except json.JSONDecodeError:
+            raise HTTPException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                detail="Server error: tagged_user data is invalid JSON",
+            )
+
+    for username in user_tag.tagged_users:
+        if username == user.get("username"):
+            raise HTTPException(status_code=400, detail="User can't tag themselves")
+
+        if query_post.is_user_tagged(username):
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail=f"User '{username}' already tagged",
+            )
+
+        tagged_user = db.query(Users).filter(Users.username == username).first()
+        if not tagged_user:
+            raise HTTPException(status_code=404, detail=f"User '{username}' not found")
+
+    query_post.tagged_user = json.dumps(user_tag.tagged_users)
+    db.commit()
+    db.refresh(post)
+
+    return {"detail": "Added user tag successfully"}
+
+
+@router.delete("/{post_id}/delete_tag", status_code=status.HTTP_200_OK)
+async def remove_user_tags(
+    user: user_dependency, db: db_dependency, post_id: int = Path(gt=0)
+):
+    if user is None:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed"
+        )
+
+    query_post = db.query(Posts).filter(Posts.id == post_id).first()
+
+    if query_post is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+        )
+
+    post = db.query(Posts).filter(Posts.owner_id == user.get("id")).first()
+
+    if post is None:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Not authorized to update this post",
+        )
+
+    if query_post.tagged_user is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Post has no user tags"
+        )
+
+    db.query(Posts).filter(Posts.id == post_id).update(
+        {Posts.tagged_user: json.dumps([])}
+    )
+    db.commit()
+
+    return {"detail": "Tagged users removed from post"}
+
+
 @router.delete("/{post_id}/delete", status_code=status.HTTP_200_OK)
 async def delete_post(
     user: user_dependency, db: db_dependency, post_id: int = Path(gt=0)
@@ -207,6 +336,14 @@ async def delete_post(
     if query_post is None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+        )
+
+    post = db.query(Posts).filter(Posts.owner_id == user.get("id")).first()
+
+    if post is None:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Not authorized to delete this post",
         )
 
     db.query(Posts).filter(Posts.id == post_id).delete()
